fix: fix 4th task

+++ 04-copy-directory/index.js
@@ -1,10 +1,26 @@
+const fs = require('fs');
 const fsPromises = require('fs/promises');
 const path = require('path');
 const mainDir = path.join(__dirname, 'files');
 const copyDir = path.join(__dirname, 'files-copy');
+let dirExists = false;
+
+function checkDirExists() {
+  fs.stat(copyDir, e => {
+    if (!e) {
+      dirExists = true;
+      mkCopyDir();
+    } else {
+      dirExists = false;
+      mkCopyDir();
+    }
+  });
+}
 
 async function mkCopyDir() {
-  await fsPromises.rmdir(copyDir, {recursive: true});
+  if (dirExists) {
+    await fsPromises.rmdir(copyDir, {recursive: true});
+  }
   await fsPromises.mkdir(copyDir, {recursive: true});
   const files = await fsPromises.readdir(mainDir, {withFileTypes: true});
   for (const file of files) {
@@ -16,4 +32,4 @@ async function mkCopyDir() {
   }
 }
 
-mkCopyDir();
+checkDirExists();
